requirements.txt
anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
deprecation==2.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
img2pdf==0.4.4
lxml==4.9.2
packaging==23.0
pdf2jpg==1.1
pikepdf==7.1.1
Pillow==9.4.0
pycodestyle==2.10.0
PyMuPDF==1.26.7
python-dotenv==0.21.1
python-telegram-bot==20.1
requests==2.32.5
rfc3986==1.5.0
sniffio==1.3.0
tomli==2.0.1
urllib3==2.6.3
